+++ bambi/models.py
@@ -730,7 +730,9 @@ class Model:
         return idata
 
     # pylint: disable=protected-access
-    def predict(self, idata, kind="mean", data=None, draws=None, inplace=True):
+    def predict(
+        self, idata, kind="mean", data=None, draws=None, inplace=True, include_group_specific=True
+    ):
         """Predict method for Bambi models
 
         Obtains in-sample and out-of-sample predictions from a fitted Bambi model.
@@ -748,6 +750,10 @@ class Model:
         data: pandas.DataFrame or None
             An optional data frame with values for the predictors that are used to obtain
             out-of-sample predictions. If omitted, the original dataset is used.
+        include_group_specific: bool
+            If ``True`` make predictions including the group specific effects. Otherwise,
+            predictions are made with common effects only (i.e. group specific are set
+            to zero).
         draws: None
             The number of random draws per chain. Only used if ``kind="pps"``. Not recommended
             unless more than ndraws times nchains posterior predictive samples are needed.
@@ -787,7 +793,7 @@ class Model:
             else:
                 X = self._design.common._evaluate_new_data(data).design_matrix
 
-        if self._design.group:
+        if self._design.group and include_group_specific:
             if in_sample:
                 Z = self._design.group.design_matrix
             else: